Flask_Project/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
gspread
google-auth
psycopg2-binary
